requirements.txt
streamlit==1.56.0
python-dotenv==1.2.2
langchain==1.2.15
langchain-groq==1.1.2
langchain-chroma==1.1.0
langchain-huggingface==1.2.2
langchain-community==0.4.1
langchain-core==1.3.1
chromadb==1.5.8
sentence-transformers==5.4.1
groq==0.37.1